split previews for SidebarView and SidebarViewController
